refactor(test): share command no-op renderer fixture
